Moved database handler to scripts folder
In hindsight, the database handler probably should have been created here in the first place.
